build: add kernel addons module

+++ build.sh
@@ -3,9 +3,12 @@
 abort()
 {
     cd -
-    echo "-----------------------------------------------"
-    echo "Kernel compilation failed! Exiting..."
-    echo "-----------------------------------------------"
+    echo "---------------------------------------------------------"
+    echo "-- Kernel compilation failed! Exiting..."
+    echo "---------------------------------------------------------"
+    if [[ "$LOCAL" == "y" ]]; then
+        FUNC_CLEANUP
+    fi
     exit -1
 }
 
@@ -33,6 +36,10 @@ while [[ $# -gt 0 ]]; do
             KERNEL_VERSION="$2"
             shift 2
             ;;
+        --rel|-r)
+            RELEASE="$2"
+            shift 2
+            ;;
         *)\
             unset_flags
             exit 1
@@ -40,9 +47,33 @@ while [[ $# -gt 0 ]]; do
     esac
 done
 
-export BUILD_CROSS_COMPILE=$(pwd)/toolchain/aarch64-linux-android-4.9/bin/aarch64-linux-androidkernel-
-export BUILD_JOB_NUMBER=`grep -c ^processor /proc/cpuinfo`
+FUNC_CHECKENV()
+{
+    DATE=`date +"%Y%m%d"`
+    export KBUILD_BUILD_USER="oItsMineZ"
+
+    if [ -n $RELEASE ]; then
+        echo "---------------------------------------------------------"
+        echo "-- Running on GitHub Actions..."
+        export KBUILD_BUILD_HOST="GitHub-Actions"
+    else
+        echo "---------------------------------------------------------"
+        echo "-- Running on Local Machine..."
+        export KBUILD_BUILD_HOST="Linux-Stable"
+        LOCAL=y
+    fi
+}
+
 RDIR=$(pwd)
+CPU=`grep -c ^processor /proc/cpuinfo`
+CLANG=$RDIR/toolchain/clang-r416183b1/bin/
+GCC=$RDIR/toolchain/aarch64-linux-android-4.9/bin/
+ARGS="
+    ARCH=arm64 O=out \
+    CC=${CLANG}clang \
+    CROSS_COMPILE=${GCC}aarch64-linux-androidkernel- \
+    CLANG_TRIPLE=${CLANG}aarch64-linux-gnu- \
+"
 
 # Define specific variables
 KERNEL_DEFCONFIG=oitsminez-"$MODEL"_defconfig
@@ -85,43 +116,47 @@ d2x)
 esac
 
 if [ -z $KSU_OPTION ]; then
-    KSU_OPTION="y"
+    KSU_OPTION=y
 fi
 
 if [[ "$KSU_OPTION" == "y" ]]; then
+    FUNC_CHECKENV
     KSU_NEXT=ksu_next.config
 
-    if test -d "$DIR/drivers/kernelsu"; then
-        echo "-----------------------------------------------"
-        echo "KernelSU-Next Directory Found!..."
-        echo "-----------------------------------------------"
+    if test -d "$DIR/drivers/kernelsu" && grep -rnw 'fs/Makefile' -e 'CONFIG_KSU_SUSFS'; then
+        echo "---------------------------------------------------------"
+        echo "-- KernelSU-Next Directory Found!..."
+        echo "---------------------------------------------------------"
     else
-        echo "-----------------------------------------------"
-        echo "Checkout oItsMineZ's KernelSU-Next Repo..."
-        echo "-----------------------------------------------"
+        echo "---------------------------------------------------------"
+        echo "-- Checkout oItsMineZ's KernelSU-Next Repo..."
+        echo "---------------------------------------------------------"
 
         git submodule add --force https://github.com/oItsMineZ/KernelSU-Next
         curl -LSs "https://raw.githubusercontent.com/oItsMineZ/KernelSU-Next/susfs-v1.5.5/kernel/setup.sh" | bash -
 
-        echo "-----------------------------------------------"
-        echo "Patch KernelSU-Next with SuSFS..."
-        echo "-----------------------------------------------"
+        if ! grep -rnw 'fs/Makefile' -e 'CONFIG_KSU_SUSFS'; then
 
-        curl -LOSs "https://raw.githubusercontent.com/oItsMineZKernel/Kernel-Patch/main/SuSFS.patch"
-        patch -p1 < SuSFS.patch
-        rm -rf *.patch
+            echo "---------------------------------------------------------"
+            echo "-- Patch KernelSU-Next with SuSFS..."
+            echo "---------------------------------------------------------"
+
+            curl -LOSs "https://raw.githubusercontent.com/oItsMineZKernel/Kernel-Patch/main/SuSFS.patch"
+            patch -p1 < SuSFS.patch
+            rm -rf *.patch
+        fi
     fi
 fi
 
-if [ -z "$KERNEL_VERSION" ]; then
+if [ -z $KERNEL_VERSION ]; then
     KERNEL_VERSION="Unofficial"
 fi
 
 FUNC_TOOLCHAIN()
 {
-    echo "-----------------------------------------------"
-    echo "Checkout Toolchain Repo..."
-    echo "-----------------------------------------------"
+    echo "---------------------------------------------------------"
+    echo "-- Checkout Toolchain Repo..."
+    echo "---------------------------------------------------------"
 
     git submodule add --branch aarch64-linux-android-4.9 --force https://github.com/oItsMineZKernel/toolchain toolchain/aarch64-linux-android-4.9
     git submodule add --branch clang-r416183b1 --force https://github.com/oItsMineZKernel/toolchain toolchain/clang-r416183b1
@@ -130,16 +165,16 @@ FUNC_TOOLCHAIN()
 FUNC_BUILD_KERNEL()
 {
     # Build kernel image
-    echo "-----------------------------------------------"
-    echo "Device: "$MODEL""
-    echo "SOC: Exynos$SOC"
-    echo "Defconfig: "$KERNEL_DEFCONFIG""
-    echo "Kernel Version: $KERNEL_VERSION"
+    echo "---------------------------------------------------------"
+    echo "-- Device: "$MODEL""
+    echo "-- SOC: Exynos$SOC"
+    echo "-- Defconfig: "$KERNEL_DEFCONFIG""
+    echo "-- Kernel Version: $KERNEL_VERSION"
 
-    if [ -z "$KSU_NEXT" ]; then
-        echo "KernelSU Next with SuSFS: Not Include"
+    if [ -z $KSU_NEXT ]; then
+        echo "-- KernelSU Next with SuSFS: Not Include"
     else
-        echo "KernelSU Next with SuSFS: $KSU_NEXT"
+        echo "-- KernelSU Next with SuSFS: $KSU_NEXT"
     fi
 
     echo -e "CONFIG_LOCALVERSION_AUTO=n" > $RDIR/arch/arm64/configs/version.config
@@ -150,28 +185,29 @@ FUNC_BUILD_KERNEL()
         DEVICE=S10
     fi
 
+    if [ -n $RELEASE ]; then
+        echo BUILD_DEVICE=$DEVICE >> $GITHUB_ENV
+    fi
+
     echo -e "CONFIG_LOCALVERSION=\"-oItsMineZKernel-"$KERNEL_VERSION"-"$DEVICE"\"" >> $RDIR/arch/arm64/configs/version.config
+    DEFCONFIG="$KERNEL_DEFCONFIG oitsminez.config version.config $KSU_NEXT"
 
-    echo "-----------------------------------------------"
-    echo "Building Kernel Using "$KERNEL_DEFCONFIG""
-    echo "Generating Configuration Files..."
-    echo "-----------------------------------------------"
+    echo "---------------------------------------------------------"
+    echo "-- Building Kernel Using "$KERNEL_DEFCONFIG""
+    echo "-- Generating Configuration Files..."
+    echo "---------------------------------------------------------"
 
-    make -j$BUILD_JOB_NUMBER ARCH=arm64 \
-        CROSS_COMPILE=$BUILD_CROSS_COMPILE O=out \
-        $KERNEL_DEFCONFIG oitsminez.config version.config $KSU_NEXT || abort
+    make -j$CPU $ARGS $DEFCONFIG || abort
 
+    echo "---------------------------------------------------------"
+    echo "-- Building Kernel..."
+    echo "---------------------------------------------------------"
 
-    echo "-----------------------------------------------"
-    echo "Building Kernel..."
-    echo "-----------------------------------------------"
+    make -j$CPU $ARGS || abort
 
-    make -j$BUILD_JOB_NUMBER ARCH=arm64 \
-        CROSS_COMPILE=$BUILD_CROSS_COMPILE O=out || abort
-
-    echo "-----------------------------------------------"
-    echo " Finished Kernel Build!"
-    echo "-----------------------------------------------"
+    echo "---------------------------------------------------------"
+    echo "-- Finished Kernel Build!"
+    echo "---------------------------------------------------------"
 
     rm -rf $RDIR/build/out/$MODEL
     mkdir -p $RDIR/build/out/$MODEL
@@ -180,17 +216,17 @@ FUNC_BUILD_KERNEL()
 FUNC_BUILD_DTBO()
 {
     # Build dtb
-    echo "Building common exynos$SOC Device Tree Blob Image..."
-    echo "-----------------------------------------------"
+    echo "-- Building common exynos$SOC Device Tree Blob Image..."
+    echo "---------------------------------------------------------"
 
     $RDIR/build/mkdtimg cfg_create $RDIR/build/out/$MODEL/dtb_exynos$SOC.img \
         $RDIR/build/dtconfigs/exynos$SOC.cfg \
         -d $RDIR/out/arch/arm64/boot/dts/exynos
 
     # Build dtbo
-    echo "-----------------------------------------------"
-    echo "Building Device Tree Blob Output Image for "$MODEL"..."
-    echo "-----------------------------------------------"
+    echo "---------------------------------------------------------"
+    echo "-- Building Device Tree Blob Output Image for "$MODEL"..."
+    echo "---------------------------------------------------------"
 
     $RDIR/build/mkdtimg cfg_create $RDIR/build/out/$MODEL/dtbo_$MODEL.img \
         $RDIR/build/dtconfigs/$MODEL.cfg \
@@ -200,17 +236,17 @@ FUNC_BUILD_DTBO()
 FUNC_BUILD_RAMDISK()
 {
     # Build ramdisk
-    echo "-----------------------------------------------"
-    echo "Building Ramdisk..."
-    echo "-----------------------------------------------"
+    echo "---------------------------------------------------------"
+    echo "-- Building Ramdisk..."
+    echo "---------------------------------------------------------"
 
     rm -f $RDIR/build/AIK/split_img/boot.img-kernel
     cp $RDIR/out/arch/arm64/boot/Image $RDIR/build/AIK/split_img/boot.img-kernel
     echo $BOARD > build/AIK/split_img/boot.img-board
 
     # Create boot image
-    echo "Creating Boot Image..."
-    echo "-----------------------------------------------"
+    echo "-- Creating Boot Image..."
+    echo "---------------------------------------------------------"
 
     # This is kinda ugly hack, we could as well touch .placeholder to all of those
     mkdir -p $RDIR/build/AIK/ramdisk/debug_ramdisk
@@ -230,12 +266,18 @@ FUNC_BUILD_RAMDISK()
 FUNC_BUILD_ZIP()
 {
     # Build zip
-    echo "-----------------------------------------------"
-    echo "Building Zip..."
+    echo "---------------------------------------------------------"
+    echo "-- Building Zip..."
+    if [[ "$LOCAL" == "y" ]] || [[ "$RELEASE" == "y" ]]; then
+        echo "---------------------------------------------------------"
+    fi
 
     rm -rf $RDIR/build/out/$MODEL/zip
     mkdir -p $RDIR/build/export
     mkdir -p $RDIR/build/out/$MODEL/zip
+    mkdir -p $RDIR/build/out/$MODEL/zip/module
+    mkdir -p $RDIR/build/out/$MODEL/zip/module/common/
+    mkdir -p $RDIR/build/out/$MODEL/zip/module/META-INF/com/google/android
     mkdir -p $RDIR/build/out/$MODEL/zip/META-INF/com/google/android
     mv $RDIR/build/AIK/image-new.img $RDIR/build/out/$MODEL/boot-patched.img
 
@@ -243,34 +285,55 @@ FUNC_BUILD_ZIP()
     cp $RDIR/build/out/$MODEL/boot-patched.img $RDIR/build/out/$MODEL/zip/boot.img
     cp $RDIR/build/out/$MODEL/dtb_exynos$SOC.img $RDIR/build/out/$MODEL/zip/dtb.img
     cp $RDIR/build/out/$MODEL/dtbo_$MODEL.img $RDIR/build/out/$MODEL/zip/dtbo.img
-    cp $RDIR/build/updater-script $RDIR/build/out/$MODEL/zip/META-INF/com/google/android/
     cp $RDIR/build/update-binary $RDIR/build/out/$MODEL/zip/META-INF/com/google/android/
-    cd $RDIR/build/out/$MODEL/zip
+    cp $RDIR/build/updater-script $RDIR/build/out/$MODEL/zip/META-INF/com/google/android/
+
+    cp $RDIR/build/module.prop $RDIR/build/out/$MODEL/zip/module/
+    cp $RDIR/build/system.prop $RDIR/build/out/$MODEL/zip/module/common/
+    cp $RDIR/build/module-binary $RDIR/build/out/$MODEL/zip/module/META-INF/com/google/android/update-binary
+    cp $RDIR/build/module-script $RDIR/build/out/$MODEL/zip/module/META-INF/com/google/android/updater-script
+
+    sed -i "s|name=oItsMineZKernel Addons for Exynos9820/9825|name=oItsMineZKernel $KERNEL_VERSION Addons for Exynos9820/9825|" $RDIR/build/out/$MODEL/zip/module/module.prop
+    sed -i "s|package_extract_file(\"module.zip\", \"/sdcard/oItsMineZKernel-Addons-.zip\");|package_extract_file(\"module.zip\", \"/sdcard/oItsMineZKernel-Addons-$KERNEL_VERSION.zip\");|" $RDIR/build/out/$MODEL/zip/META-INF/com/google/android/updater-script
+    sed -i "s|ui_print(\"   /sdcard/oItsMineZKernel-Addons-.zip\");|ui_print(\"   /sdcard/oItsMineZKernel-Addons-$KERNEL_VERSION.zip\");|" $RDIR/build/out/$MODEL/zip/META-INF/com/google/android/updater-script
+
+    cd $RDIR/build/out/$MODEL/zip/module
+    zip -r ../module.zip .
+    rm -rf $RDIR/build/out/$MODEL/zip/module
 
     sed -i "s/ui_print(\" Kernel Version: \");/ui_print(\" Kernel Version: $KERNEL_VERSION\");/" $RDIR/build/out/$MODEL/zip/META-INF/com/google/android/updater-script
     sed -i "s/ui_print(\" Kernel Device: \");/ui_print(\" Kernel Device: $DEVICE ($MODEL)\");/" $RDIR/build/out/$MODEL/zip/META-INF/com/google/android/updater-script
+    sed -i "s/CONFIG_LOCALVERSION=\"-oItsMineZKernel-"$KERNEL_VERSION"-"$DEVICE"\"/CONFIG_LOCALVERSION=\"-oItsMineZKernel-"$KERNEL_VERSION"-"$DATE"-"$DEVICE"\"/" $RDIR/arch/arm64/configs/version.config
 
-    version=$(grep -o 'CONFIG_LOCALVERSION="[^"]*"' "$RDIR"/arch/arm64/configs/version.config | cut -d '"' -f 2)
+    version=$(grep -o 'CONFIG_LOCALVERSION="[^"]*"' $RDIR/arch/arm64/configs/version.config | cut -d '"' -f 2)
     version=${version:1}
-
-    DATE=`date +"%d-%m-%Y"`    
     NAME="$version"-"$MODEL".zip
 
-    zip -r ../"$NAME" .
-    rm -rf $RDIR/build/out/$MODEL/zip
-    mv $RDIR/build/out/$MODEL/"$NAME" $RDIR/build/export/"$NAME"
-    cd $RDIR/build/export
+    if [[ "$LOCAL" == "y" ]] || [[ "$RELEASE" == "y" ]]; then
+        cd $RDIR/build/out/$MODEL/zip
+        zip -r ../"$NAME" .
+        rm -rf $RDIR/build/out/$MODEL/zip
+        mv $RDIR/build/out/$MODEL/"$NAME" $RDIR/build/export/"$NAME"
+        cd $RDIR/build/export
+    fi
 }
 
 FUNC_CLEANUP()
 {
-    echo "-----------------------------------------------"
-    echo "Cleanup build files..."
-    echo "-----------------------------------------------"
+    echo "---------------------------------------------------------"
+    echo "-- Cleanup build files..."
+    echo "---------------------------------------------------------"
 
-    rm -rf $RDIR/build/AIK/ramdisk/fstab.exynos*
-    rm -rf $RDIR/arch/arm64/configs/version.config
-    git reset $RDIR/build/AIK/split_img/boot.img-board
+    cd $RDIR && rm -rf out
+    rm -rf .wireguard-fetch-lock
+    rm -rf arch/arm64/configs/version.config
+    rm -rf build/AIK/ramdisk/fstab.exynos*
+    rm -rf build/AIK/split_img/boot.img-kernel
+
+    git rm -rf KernelSU-Next
+    git rm -rf toolchain
+
+    git reset --hard HEAD && git clean -df
 }
 
 # MAIN FUNCTION
@@ -278,13 +341,13 @@ rm -rf ./build.log
 (
     START_TIME=`date +%s`
 
-    echo "-----------------------------------------------"
-    echo "Preparing the Build Environment..."
+    echo "---------------------------------------------------------"
+    echo "-- Preparing the Build Environment..."
 
     if test -d "$DIR/toolchain"; then
-        echo "-----------------------------------------------"
-        echo "Toolchain Directory Found!"
-        echo "-----------------------------------------------"
+        echo "---------------------------------------------------------"
+        echo "-- Toolchain Directory Found!"
+        echo "---------------------------------------------------------"
     else
         FUNC_TOOLCHAIN
     fi
@@ -293,13 +356,16 @@ rm -rf ./build.log
     FUNC_BUILD_DTBO
     FUNC_BUILD_RAMDISK
     FUNC_BUILD_ZIP
-    FUNC_CLEANUP
+
+    if [[ "$LOCAL" == "y" ]]; then
+        FUNC_CLEANUP
+    fi
 
     END_TIME=`date +%s`
 
     let "ELAPSED_TIME=$END_TIME-$START_TIME"
 
-    echo "-----------------------------------------------"
-    echo "Total compile time was $ELAPSED_TIME seconds"
-    echo "-----------------------------------------------"
+    echo "---------------------------------------------------------"
+    echo "-- Total compile time was $(($ELAPSED_TIME / 60)) minutes and $(($ELAPSED_TIME % 60)) seconds"
+    echo "---------------------------------------------------------"
 ) 2>&1	| tee -a ./build.log
